Fixed flicker with host-coherent buffers...
Turned out that synching between host-writes and GPU-usage of the
buffers was NOT the problem, but the previous command buffer for the
same frame-in-flight-index was still executing.
Waiting for its fence before updating the buffers fixed the problem.
(Most likely this was also the source of the "beta driver crash"
encountered earlier.)